langchain/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
openai>=1.58.0
tiktoken>=0.8.0
python-dotenv>=1.0.0
